+++ src/Interfaces/FrameInterface.php
@@ -6,8 +6,8 @@ interface FrameInterface
 {
     public function toImage(): ImageInterface;
     public function getCore();
-    public function getDelay(): int;
-    public function setDelay(int $delay): FrameInterface;
+    public function getDelay(): float;
+    public function setDelay(float $delay): FrameInterface;
     public function getDispose(): int;
     public function setDispose(int $dispose): FrameInterface;
     public function setOffset(int $left, int $top): FrameInterface;
